+++ tgpt2/__init__.py
@@ -1,7 +1,7 @@
 from .tgpt2 import TGPT
 from .imager import Imager
 
-__version__ = "0.0.1"
+__version__ = "0.0.2"
 __author__ = "Smartwa"
 __repo__ = "https://github.com/Simatwa/tgpt2"