+++ docs/CHANGELOG.md
@@ -3,3 +3,11 @@
 **What's new?**
 
 - Initial release. 
+
+## v0.0.2
+
+**What's new?**
+- Enhanced response generation
+- Prompt optimizers added - *[code, shell_command]*
+- Prompt optimizers added - *Console*
+- Clear console  - *Console*